Unused Files removed and session management working on client side
Session is managed using sessionStorage
, access token is stored with accessToken.
When user is logged in (there is an access token in sessionStorage), hide login component and showing logout component and vice versa.
when logging out, session on server is not deleted, nor there is a provision to refresh the access token, it is valid for 1 hour and will work until then.