Added functionality to log fileactions from within the app (work in progress see description)
Currently only logging dummy data when pressing the "do something" option on the dropdown menu in the file explorer. Next steps are to capture actual user/file data, and display the logs somewhere.

If there is no database connection available clicking the button will just fail in the background without affecting the rest of the app.