FAQ improvements/corrections
Removed and corrected false and misleading information in the FAQ
Decomposed the complexity of the original FAQ, breaking down questions into individual componets in-line with React recommendations and practices.

In the future, this approach will also allow the FAQ questions to become multilingual, switching the language state of the FAQ, can toggle access to which language the individual questions may need to be.